Chat/api/requirements.txt
grpcio
grpcio-tools
fastapi
uvicorn
kafka-python
cassandra-driver
httpx
PyJWT
python-multipart 
requests
minio
prometheus-fastapi-instrumentator 
prometheus-client
python-telegram-bot==20.7
python-dotenv
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-logging
python-json-logger
